requirements.txt
Django==5.0.2
python-dateutil
